MyApp/requirements.txt
Flask==3.1.2
Werkzeug==3.1.2
Pillow==10.1.0
